+++ .agents/skills/remix/SKILL.md
@@ -431,24 +431,24 @@ what it exports. Open the linked reference file when you need full examples.
 ### Define routes first
 
 ```typescript
-import { form, get, post, resources, route } from 'remix/routes'
+import { form, get, post, resources, route } from "remix/routes";
 
 export const routes = route({
-  home: '/',
-  contact: form('contact'),
+  home: "/",
+  contact: form("contact"),
   books: {
-    index: '/books',
-    show: '/books/:slug',
+    index: "/books",
+    show: "/books/:slug",
   },
-  auth: route('auth', {
-    login: form('login'),
-    logout: post('logout'),
+  auth: route("auth", {
+    login: form("login"),
+    logout: post("logout"),
   }),
-  admin: route('admin', {
-    index: get('/'),
-    books: resources('books', { param: 'bookId' }),
+  admin: route("admin", {
+    index: get("/"),
+    books: resources("books", { param: "bookId" }),
   }),
-})
+});
 ```
 
 ### Type controllers against the route contract
@@ -478,49 +478,49 @@ export default createController(routes.books, {
 ### Register Controllers Explicitly
 
 ```typescript
-import { createRouter } from 'remix/router'
+import { createRouter } from "remix/router";
 
-import rootController from './actions/controller.tsx'
-import adminController from './actions/admin/controller.tsx'
-import adminBooksController from './actions/admin/books/controller.tsx'
-import authController from './actions/auth/controller.tsx'
-import authLoginController from './actions/auth/login/controller.tsx'
-import booksController from './actions/books/controller.tsx'
-import contactController from './actions/contact/controller.tsx'
-import { routes } from './routes.ts'
+import rootController from "./actions/controller.tsx";
+import adminController from "./actions/admin/controller.tsx";
+import adminBooksController from "./actions/admin/books/controller.tsx";
+import authController from "./actions/auth/controller.tsx";
+import authLoginController from "./actions/auth/login/controller.tsx";
+import booksController from "./actions/books/controller.tsx";
+import contactController from "./actions/contact/controller.tsx";
+import { routes } from "./routes.ts";
 
-export const router = createRouter({ middleware })
+export const router = createRouter({ middleware });
 
-router.map(routes, rootController)
-router.map(routes.contact, contactController)
-router.map(routes.books, booksController)
-router.map(routes.auth, authController)
-router.map(routes.auth.login, authLoginController)
-router.map(routes.admin, adminController)
-router.map(routes.admin.books, adminBooksController)
+router.map(routes, rootController);
+router.map(routes.contact, contactController);
+router.map(routes.books, booksController);
+router.map(routes.auth, authController);
+router.map(routes.auth.login, authLoginController);
+router.map(routes.admin, adminController);
+router.map(routes.admin.books, adminBooksController);
 ```
 
 ### Compose middleware deliberately
 
 ```typescript
-import { createRouter } from 'remix/router'
+import { createRouter } from "remix/router";
 
-let middleware = []
+let middleware = [];
 
-if (process.env.NODE_ENV === 'development') {
-  middleware.push(logger())
+if (process.env.NODE_ENV === "development") {
+  middleware.push(logger());
 }
 
-middleware.push(compression())
-middleware.push(staticFiles('./public'))
-middleware.push(formData())
-middleware.push(methodOverride())
-middleware.push(session(cookie, storage))
-middleware.push(asyncContext())
-middleware.push(loadDatabase())
-middleware.push(loadAuth())
+middleware.push(compression());
+middleware.push(staticFiles("./public"));
+middleware.push(formData());
+middleware.push(methodOverride());
+middleware.push(session(cookie, storage));
+middleware.push(asyncContext());
+middleware.push(loadDatabase());
+middleware.push(loadAuth());
 
-let router = createRouter({ middleware })
+let router = createRouter({ middleware });
 ```
 
 ### Validate, mutate, and respond
@@ -566,21 +566,21 @@ This shape works without JavaScript, returns a `Response` for every outcome, and
 ### Build UI from handle props plus render
 
 ```tsx
-import { on, type Handle } from 'remix/ui'
+import { on, type Handle } from "remix/ui";
 
 function Counter(handle: Handle<{ initialCount?: number; label: string }>) {
-  let count = handle.props.initialCount ?? 0
+  let count = handle.props.initialCount ?? 0;
 
   return () => (
     <button
-      mix={on('click', () => {
-        count++
-        handle.update()
+      mix={on("click", () => {
+        count++;
+        handle.update();
       })}
     >
       {handle.props.label}: {count}
     </button>
-  )
+  );
 }
 ```
