+++ .agents/skills/remix/references/animate-elements.md
@@ -24,8 +24,8 @@ Animates an element when inserted. Config specifies the **starting** style the e
 <div
   mix={animateEntrance({
     opacity: 0,
-    transform: 'translateY(8px)',
-    ...spring('smooth'),
+    transform: "translateY(8px)",
+    ...spring("smooth"),
   })}
 />
 ```
@@ -41,11 +41,15 @@ Animates an element when removed. Config specifies the **ending** style the elem
     <div
       key="panel"
       mix={[
-        animateEntrance({ opacity: 0, transform: 'scale(0.98)', ...spring('smooth') }),
-        animateExit({ opacity: 0, duration: 120, easing: 'ease-in' }),
+        animateEntrance({
+          opacity: 0,
+          transform: "scale(0.98)",
+          ...spring("smooth"),
+        }),
+        animateExit({ opacity: 0, duration: 120, easing: "ease-in" }),
       ]}
     />
-  )
+  );
 }
 ```
 
@@ -56,8 +60,11 @@ Animates layout changes (position/size) using FLIP-style transforms:
 ```tsx
 {
   items.map((item) => (
-    <li key={item.id} mix={animateLayout({ ...spring({ duration: 500, bounce: 0.2 }) })} />
-  ))
+    <li
+      key={item.id}
+      mix={animateLayout({ ...spring({ duration: 500, bounce: 0.2 }) })}
+    />
+  ));
 }
 ```
 
@@ -70,9 +77,18 @@ include scale projection for size changes).
 <div
   key="card"
   mix={[
-    animateEntrance({ opacity: 0, transform: 'scale(0.95)', ...spring('snappy') }),
-    animateExit({ opacity: 0, transform: 'scale(0.98)', duration: 120, easing: 'ease-in' }),
-    animateLayout({ duration: 220, easing: 'ease-out' }),
+    animateEntrance({
+      opacity: 0,
+      transform: "scale(0.95)",
+      ...spring("snappy"),
+    }),
+    animateExit({
+      opacity: 0,
+      transform: "scale(0.98)",
+      duration: 120,
+      easing: "ease-in",
+    }),
+    animateLayout({ duration: 220, easing: "ease-out" }),
   ]}
 />
 ```
@@ -80,11 +96,17 @@ include scale projection for size changes).
 ### Shared-layout swap
 
 ```tsx
-<div mix={css({ display: 'grid', '& > *': { gridArea: '1 / 1' } })}>
+<div mix={css({ display: "grid", "& > *": { gridArea: "1 / 1" } })}>
   {stateA ? (
-    <div key="a" mix={[animateEntrance({ opacity: 0 }), animateExit({ opacity: 0 })]} />
+    <div
+      key="a"
+      mix={[animateEntrance({ opacity: 0 }), animateExit({ opacity: 0 })]}
+    />
   ) : (
-    <div key="b" mix={[animateEntrance({ opacity: 0 }), animateExit({ opacity: 0 })]} />
+    <div
+      key="b"
+      mix={[animateEntrance({ opacity: 0 }), animateExit({ opacity: 0 })]}
+    />
   )}
 </div>
 ```
@@ -103,17 +125,17 @@ Physics-based spring animation. Returns a `SpringIterator` with `duration`, `eas
 | `bouncy` | 0.3    | 400ms    | Underdamped, visible bounce |
 
 ```tsx
-spring('bouncy')
-spring('snappy')
-spring('smooth')
-spring('bouncy', { duration: 300 }) // override duration
+spring("bouncy");
+spring("snappy");
+spring("smooth");
+spring("bouncy", { duration: 300 }); // override duration
 ```
 
 ### Custom spring
 
 ```tsx
-spring({ duration: 500, bounce: 0.3 })
-spring({ duration: 500, bounce: 0.3, velocity: 2 }) // continue momentum from gesture
+spring({ duration: 500, bounce: 0.3 });
+spring({ duration: 500, bounce: 0.3, velocity: 2 }); // continue momentum from gesture
 ```
 
 ### Spread into animation mixins
@@ -121,7 +143,7 @@ spring({ duration: 500, bounce: 0.3, velocity: 2 }) // continue momentum from ge
 Spreading a spring gives both `duration` and `easing`:
 
 ```tsx
-animateEntrance({ opacity: 0, ...spring('bouncy') })
+animateEntrance({ opacity: 0, ...spring("bouncy") });
 ```
 
 ### CSS transitions
@@ -129,20 +151,20 @@ animateEntrance({ opacity: 0, ...spring('bouncy') })
 The iterator stringifies to `"550ms linear(...)"`:
 
 ```tsx
-css({ transition: `width ${spring('bouncy')}` })
+css({ transition: `width ${spring("bouncy")}` });
 ```
 
 Or use the `spring.transition()` helper for multiple properties:
 
 ```tsx
-css({ transition: spring.transition('width', 'bouncy') })
-css({ transition: spring.transition(['left', 'top'], 'snappy') })
+css({ transition: spring.transition("width", "bouncy") });
+css({ transition: spring.transition(["left", "top"], "snappy") });
 ```
 
 ### Web Animations API
 
 ```tsx
-element.animate(keyframes, { ...spring('bouncy') })
+element.animate(keyframes, { ...spring("bouncy") });
 ```
 
 ### JS iteration
@@ -150,10 +172,10 @@ element.animate(keyframes, { ...spring('bouncy') })
 The iterator yields position values from 0 to 1, one per frame:
 
 ```tsx
-for (let t of spring('bouncy')) {
-  let x = from + (to - from) * t
-  updateSomething(x)
-  await nextFrame()
+for (let t of spring("bouncy")) {
+  let x = from + (to - from) * t;
+  updateSomething(x);
+  await nextFrame();
 }
 ```
 
@@ -164,23 +186,23 @@ mixins or CSS transitions with `spring` for most UI work. Use `tween` for impera
 `requestAnimationFrame` loops, canvas/WebGL, or non-CSS properties.
 
 ```tsx
-import { tween, easings } from 'remix/ui/animation'
+import { tween, easings } from "remix/ui/animation";
 
 let animation = tween({
   from: 0,
   to: 100,
   duration: 300,
   curve: easings.easeOut,
-})
+});
 
-animation.next() // initialize
+animation.next(); // initialize
 function tick(timestamp: number) {
-  if (handle.signal.aborted) return
-  let { value, done } = animation.next(timestamp)
-  element.style.transform = `translateX(${value}px)`
-  if (!done) requestAnimationFrame(tick)
+  if (handle.signal.aborted) return;
+  let { value, done } = animation.next(timestamp);
+  element.style.transform = `translateX(${value}px)`;
+  if (!done) requestAnimationFrame(tick);
 }
-requestAnimationFrame(tick)
+requestAnimationFrame(tick);
 ```
 
 Built-in easings: `easings.linear`, `easings.ease`, `easings.easeIn`, `easings.easeOut`,
