+++ .agents/skills/remix/references/create-mixins.md
@@ -28,24 +28,24 @@ into one semantic event, or when the pattern is reused across components.
    work.
 
 ```tsx
-import { createMixin } from 'remix/ui'
+import { createMixin } from "remix/ui";
 
 let myMixin = createMixin<HTMLElement>((handle) => {
-  handle.addEventListener('insert', (event) => {
+  handle.addEventListener("insert", (event) => {
     // event.node is the mounted host node
-  })
+  });
 
-  handle.addEventListener('remove', () => {
+  handle.addEventListener("remove", () => {
     // Clean up listeners, timers, observers
-  })
+  });
 
   return (props) => {
     handle.queueTask((node) => {
       // Post-commit work that needs the concrete host node
-    })
-    return <handle.element {...props} />
-  }
-})
+    });
+    return <handle.element {...props} />;
+  };
+});
 ```
 
 ## Patterns
@@ -53,20 +53,22 @@ let myMixin = createMixin<HTMLElement>((handle) => {
 ### Pure prop transform
 
 ```tsx
-let withTitle = createMixin((handle) => (title: string, props: { title?: string }) => (
-  <handle.element {...props} title={title} />
-))
+let withTitle = createMixin(
+  (handle) => (title: string, props: { title?: string }) => (
+    <handle.element {...props} title={title} />
+  ),
+);
 ```
 
 ### Lifecycle-managed imperative setup
 
 ```tsx
 let withFocus = createMixin<HTMLElement>((handle) => {
-  handle.addEventListener('insert', (event) => {
-    event.node.focus()
-  })
-  return (props) => <handle.element {...props} />
-})
+  handle.addEventListener("insert", (event) => {
+    event.node.focus();
+  });
+  return (props) => <handle.element {...props} />;
+});
 ```
 
 ## Custom Event Mixins
@@ -80,68 +82,68 @@ is reused across components.
 4. Dispatch from the host node inside the mixin.
 
 ```tsx
-import { createMixin, on } from 'remix/ui'
+import { createMixin, on } from "remix/ui";
 
-export let dragReleaseType = 'myapp:drag-release' as const
+export let dragReleaseType = "myapp:drag-release" as const;
 
 declare global {
   interface HTMLElementEventMap {
-    [dragReleaseType]: DragReleaseEvent
+    [dragReleaseType]: DragReleaseEvent;
   }
 }
 
 export class DragReleaseEvent extends Event {
-  velocityX: number
-  velocityY: number
+  velocityX: number;
+  velocityY: number;
   constructor(init: { velocityX: number; velocityY: number }) {
-    super(dragReleaseType, { bubbles: true, cancelable: true })
-    this.velocityX = init.velocityX
-    this.velocityY = init.velocityY
+    super(dragReleaseType, { bubbles: true, cancelable: true });
+    this.velocityX = init.velocityX;
+    this.velocityY = init.velocityY;
   }
 }
 
 export let dragRelease = createMixin<HTMLElement>((handle) => {
-  let node: HTMLElement | undefined
-  let tracking = false
-  let velocityX = 0
-  let velocityY = 0
-  let lastX = 0
-  let lastY = 0
-  let lastT = 0
+  let node: HTMLElement | undefined;
+  let tracking = false;
+  let velocityX = 0;
+  let velocityY = 0;
+  let lastX = 0;
+  let lastY = 0;
+  let lastT = 0;
 
-  handle.addEventListener('insert', (event) => {
-    node = event.node
-  })
+  handle.addEventListener("insert", (event) => {
+    node = event.node;
+  });
 
   return () => (
     <handle.element
       mix={[
-        on('pointerdown', (event) => {
-          if (!event.isPrimary) return
-          tracking = true
-          lastX = event.clientX
-          lastY = event.clientY
-          lastT = event.timeStamp
-          node?.setPointerCapture(event.pointerId)
+        on("pointerdown", (event) => {
+          if (!event.isPrimary) return;
+          tracking = true;
+          lastX = event.clientX;
+          lastY = event.clientY;
+          lastT = event.timeStamp;
+          node?.setPointerCapture(event.pointerId);
         }),
-        on('pointermove', (event) => {
-          if (!tracking) return
-          let dt = Math.max(1, event.timeStamp - lastT)
-          velocityX = (event.clientX - lastX) / dt
-          velocityY = (event.clientY - lastY) / dt
-          lastX = event.clientX
-          lastY = event.clientY
-          lastT = event.timeStamp
+        on("pointermove", (event) => {
+          if (!tracking) return;
+          let dt = Math.max(1, event.timeStamp - lastT);
+          velocityX = (event.clientX - lastX) / dt;
+          velocityY = (event.clientY - lastY) / dt;
+          lastX = event.clientX;
+          lastY = event.clientY;
+          lastT = event.timeStamp;
         }),
-        on('pointerup', () => {
-          if (!tracking) return
-          tracking = false
-          node?.dispatchEvent(new DragReleaseEvent({ velocityX, velocityY }))
+        on("pointerup", () => {
+          if (!tracking) return;
+          tracking = false;
+          node?.dispatchEvent(new DragReleaseEvent({ velocityX, velocityY }));
         }),
       ]}
     />
-  )
-})
+  );
+});
 ```
 
 Consume it:
@@ -151,7 +153,7 @@ Consume it:
   mix={[
     dragRelease(),
     on(dragReleaseType, (event) => {
-      console.log('velocity:', event.velocityX, event.velocityY)
+      console.log("velocity:", event.velocityX, event.velocityY);
     }),
   ]}
 />
