+++ .agents/skills/remix/references/hydration-frames-navigation.md
@@ -27,16 +27,16 @@ maps cleanly to a server-rendered route. Frames re-fetch the same route, so the 
 stays in one place and the client does not need a parallel "state" API.
 
 ```tsx
-on('submit', async (event, signal) => {
-  event.preventDefault()
+on("submit", async (event, signal) => {
+  event.preventDefault();
   await fetch(routes.cart.add.href(), {
-    method: 'POST',
+    method: "POST",
     body: new FormData(event.currentTarget),
     signal,
-  })
-  if (signal.aborted) return
-  await handle.frames.get('cart-summary')?.reload()
-})
+  });
+  if (signal.aborted) return;
+  await handle.frames.get("cart-summary")?.reload();
+});
 ```
 
 Use polling or a small JSON state endpoint when the data changes outside this page, or when a tiny
@@ -50,12 +50,12 @@ source module's `import.meta.url` as the entry ID and let server rendering map i
 asset URL:
 
 ```tsx
-import { clientEntry, on, type Handle } from 'remix/ui'
+import { clientEntry, on, type Handle } from "remix/ui";
 
 export const Counter = clientEntry(
   import.meta.url,
   function Counter(handle: Handle<{ initialCount: number; label: string }>) {
-    let count = handle.props.initialCount
+    let count = handle.props.initialCount;
 
     return () => (
       <div>
@@ -63,17 +63,17 @@ export const Counter = clientEntry(
           {handle.props.label}: {count}
         </span>
         <button
-          mix={on('click', () => {
-            count++
-            handle.update()
+          mix={on("click", () => {
+            count++;
+            handle.update();
           })}
         >
           +
         </button>
       </div>
-    )
+    );
   },
-)
+);
 ```
 
 On the server, provide `resolveClientEntry` to `renderToStream(...)` so source file URLs become
@@ -83,17 +83,17 @@ hard-code deployment-specific asset paths:
 ```tsx
 let stream = renderToStream(<App />, {
   async resolveClientEntry(entryId, component) {
-    let exportName = entryId.split('#')[1] || component.name
+    let exportName = entryId.split("#")[1] || component.name;
     if (!exportName) {
-      throw new Error(`Unable to resolve client entry export for ${entryId}`)
+      throw new Error(`Unable to resolve client entry export for ${entryId}`);
     }
 
     return {
       href: await assetServer.getHref(entryId),
       exportName,
-    }
+    };
   },
-})
+});
 ```
 
 If the module export name differs from the component function name, include `#ExportName` in the
@@ -113,26 +113,26 @@ Use `run` to start the client runtime. It scans the document for client entry ma
 modules, and hydrates each one:
 
 ```tsx
-import { run } from 'remix/ui'
+import { run } from "remix/ui";
 
 let app = run({
   async loadModule(moduleUrl, exportName) {
-    let mod = await import(moduleUrl)
-    return mod[exportName]
+    let mod = await import(moduleUrl);
+    return mod[exportName];
   },
   async resolveFrame(src, signal, target) {
-    let headers = new Headers({ accept: 'text/html' })
-    if (target) headers.set('x-remix-target', target)
-    let response = await fetch(src, { headers, signal })
-    return response.body ?? (await response.text())
+    let headers = new Headers({ accept: "text/html" });
+    if (target) headers.set("x-remix-target", target);
+    let response = await fetch(src, { headers, signal });
+    return response.body ?? (await response.text());
   },
-})
+});
 
-app.addEventListener('error', (event) => {
-  console.error('Component error:', event.error)
-})
+app.addEventListener("error", (event) => {
+  console.error("Component error:", event.error);
+});
 
-await app.ready()
+await app.ready();
 ```
 
 ### `run` options
@@ -156,7 +156,7 @@ A `<Frame>` renders server content into the page. Frames stream after the initia
 other frames, contain client entries, and can be reloaded without full page navigation.
 
 ```tsx
-import { Frame } from 'remix/ui'
+import { Frame } from "remix/ui";
 
 function App() {
   return () => (
@@ -164,7 +164,7 @@ function App() {
       <Frame src="/sidebar" fallback={<div>Loading...</div>} />
       <Frame name="main" src="/main-content" />
     </div>
-  )
+  );
 }
 ```
 
@@ -188,13 +188,13 @@ Client entries inside a frame can trigger a reload:
 
 ```tsx
 // Reload the containing frame
-handle.frame.reload()
+handle.frame.reload();
 
 // Reload an adjacent named frame
-await handle.frames.get('cart-summary')?.reload()
+await handle.frames.get("cart-summary")?.reload();
 
 // Reload the entire page/frame tree
-handle.frames.top.reload()
+handle.frames.top.reload();
 ```
 
 When a frame reloads, matching DOM nodes are updated in place. Client entries receive updated props
@@ -214,22 +214,22 @@ Renders a component tree to a `ReadableStream<Uint8Array>`. Sends initial HTML i
 streams frame content as it resolves:
 
 ```tsx
-import { renderToStream } from 'remix/ui/server'
+import { renderToStream } from "remix/ui/server";
 
 let stream = renderToStream(<App />, {
   frameSrc: request.url,
   resolveFrame(src, target, context) {
-    let frameUrl = new URL(src, context?.currentFrameSrc ?? request.url)
-    return fetchHtml(frameUrl)
+    let frameUrl = new URL(src, context?.currentFrameSrc ?? request.url);
+    return fetchHtml(frameUrl);
   },
   onError(error) {
-    console.error(error)
+    console.error(error);
   },
-})
+});
 
 return new Response(stream, {
-  headers: { 'Content-Type': 'text/html; charset=utf-8' },
-})
+  headers: { "Content-Type": "text/html; charset=utf-8" },
+});
 ```
 
 Options:
@@ -247,8 +247,8 @@ Options:
 Renders a component tree to a complete HTML string. Use for static pages or embedding HTML:
 
 ```tsx
-import { renderToString } from 'remix/ui/server'
-let html = await renderToString(<App />)
+import { renderToString } from "remix/ui/server";
+let html = await renderToString(<App />);
 ```
 
 ### CSS in SSR
@@ -264,8 +264,8 @@ Use real anchors for normal document navigation. For app-driven navigation:
 - `link(href, options?)` mixin — makes any element behave like a navigation link
 
 ```tsx
-import { navigate } from 'remix/ui'
-navigate('/dashboard', { history: 'replace' })
+import { navigate } from "remix/ui";
+navigate("/dashboard", { history: "replace" });
 ```
 
 Options: `src`, `target`, `history` (`'push' | 'replace'`), `resetScroll`.
@@ -289,7 +289,7 @@ function App() {
         <main>...</main>
       </body>
     </html>
-  )
+  );
 }
 ```
