+++ .agents/skills/remix/references/middleware-and-server.md
@@ -22,31 +22,31 @@ request-enriching middleware (session, auth) later.
 Recommended ordering:
 
 ```typescript
-import { createRouter } from 'remix/router'
-import { compression } from 'remix/middleware/compression'
-import { formData } from 'remix/middleware/form-data'
-import { logger } from 'remix/middleware/logger'
-import { methodOverride } from 'remix/middleware/method-override'
-import { session } from 'remix/middleware/session'
-import { staticFiles } from 'remix/middleware/static'
-import { asyncContext } from 'remix/middleware/async-context'
+import { createRouter } from "remix/router";
+import { compression } from "remix/middleware/compression";
+import { formData } from "remix/middleware/form-data";
+import { logger } from "remix/middleware/logger";
+import { methodOverride } from "remix/middleware/method-override";
+import { session } from "remix/middleware/session";
+import { staticFiles } from "remix/middleware/static";
+import { asyncContext } from "remix/middleware/async-context";
 
-let middleware = []
+let middleware = [];
 
-if (process.env.NODE_ENV === 'development') {
-  middleware.push(logger())
+if (process.env.NODE_ENV === "development") {
+  middleware.push(logger());
 }
 
-middleware.push(compression())
-middleware.push(staticFiles('./public'))
-middleware.push(formData())
-middleware.push(methodOverride())
-middleware.push(session(cookie, storage))
-middleware.push(asyncContext())
-middleware.push(loadDatabase())
-middleware.push(loadAuth())
+middleware.push(compression());
+middleware.push(staticFiles("./public"));
+middleware.push(formData());
+middleware.push(methodOverride());
+middleware.push(session(cookie, storage));
+middleware.push(asyncContext());
+middleware.push(loadDatabase());
+middleware.push(loadAuth());
 
-let router = createRouter({ middleware })
+let router = createRouter({ middleware });
 ```
 
 ### Built-in middleware catalog
@@ -96,23 +96,23 @@ let router = createRouter({ middleware })
 
 ```typescript
 // Static files with cache headers
-staticFiles('./public', {
-  cacheControl: 'no-store, must-revalidate',
+staticFiles("./public", {
+  cacheControl: "no-store, must-revalidate",
   etag: false,
   lastModified: false,
-})
+});
 
 // Form data with upload handler
-import type { FileUpload } from 'remix/form-data-parser'
-import { createFsFileStorage } from 'remix/file-storage/fs'
+import type { FileUpload } from "remix/form-data-parser";
+import { createFsFileStorage } from "remix/file-storage/fs";
 
-let fileStorage = createFsFileStorage('./tmp/uploads')
+let fileStorage = createFsFileStorage("./tmp/uploads");
 
 formData({
   uploadHandler(fileUpload: FileUpload) {
-    return fileStorage.set(fileUpload.name, fileUpload)
+    return fileStorage.set(fileUpload.name, fileUpload);
   },
-})
+});
 ```
 
 Errors thrown or rejected by `uploadHandler` propagate directly. Catch domain-specific upload
@@ -129,30 +129,30 @@ context and want the router to continue automatically.
 Use `context.set(key, value)` to add typed values accessible downstream via `context.get(key)`.
 
 ```typescript
-import type { Middleware } from 'remix/router'
-import { Database } from 'remix/data-table'
+import type { Middleware } from "remix/router";
+import { Database } from "remix/data-table";
 
 export function loadDatabase(): Middleware {
   return async (context, next) => {
-    context.set(Database, db)
-    return next()
-  }
+    context.set(Database, db);
+    return next();
+  };
 }
 ```
 
 ### Guarding routes
 
 ```typescript
-import { Auth } from 'remix/middleware/auth'
+import { Auth } from "remix/middleware/auth";
 
 export function requireAdmin(): Middleware {
   return (context, next) => {
-    let auth = context.get(Auth)
-    if (auth.identity?.role !== 'admin') {
-      return new Response('Forbidden', { status: 403 })
+    let auth = context.get(Auth);
+    if (auth.identity?.role !== "admin") {
+      return new Response("Forbidden", { status: 403 });
     }
-    return next()
-  }
+    return next();
+  };
 }
 ```
 
@@ -164,30 +164,32 @@ helpers:
 
 ```typescript
 // app/utils/context.ts
-import { getContext } from 'remix/middleware/async-context'
-import { Auth } from 'remix/middleware/auth'
-import { Database } from 'remix/data-table'
-import { Session } from 'remix/session'
+import { getContext } from "remix/middleware/async-context";
+import { Auth } from "remix/middleware/auth";
+import { Database } from "remix/data-table";
+import { Session } from "remix/session";
 
 export function getCurrentDb() {
-  return getContext().get(Database)
+  return getContext().get(Database);
 }
 
 export function getCurrentSession() {
-  return getContext().get(Session)
+  return getContext().get(Session);
 }
 
 export function getCurrentUser() {
-  let auth = getContext().get(Auth)
+  let auth = getContext().get(Auth);
   if (!auth.ok) {
-    throw new Error('Expected an authenticated user. Run requireAuth() before this code.')
+    throw new Error(
+      "Expected an authenticated user. Run requireAuth() before this code.",
+    );
   }
-  return auth.identity
+  return auth.identity;
 }
 
 export function getCurrentUserSafely() {
-  let auth = getContext().get(Auth)
-  return auth.ok ? auth.identity : null
+  let auth = getContext().get(Auth);
+  return auth.ok ? auth.identity : null;
 }
 ```
 
@@ -218,7 +220,7 @@ Middleware can be applied at three levels:
    router.get(routes.account.index, {
      middleware: [requireAuth()],
      handler: accountAction.handler,
-   })
+   });
    ```
 
 ## Node Server Setup
