+++ .agents/skills/remix/references/mixins-styling-events.md
@@ -27,18 +27,18 @@ when the handler is re-entered or the component is removed — this prevents rac
 
 ```tsx
 <input
-  mix={on('input', async (event, signal) => {
-    let query = event.currentTarget.value
-    loading = true
-    handle.update()
+  mix={on("input", async (event, signal) => {
+    let query = event.currentTarget.value;
+    loading = true;
+    handle.update();
 
-    let response = await fetch(`/search?q=${query}`, { signal })
-    let data = await response.json()
-    if (signal.aborted) return
+    let response = await fetch(`/search?q=${query}`, { signal });
+    let data = await response.json();
+    if (signal.aborted) return;
 
-    results = data.results
-    loading = false
-    handle.update()
+    results = data.results;
+    loading = false;
+    handle.update();
   })}
 />
 ```
@@ -63,17 +63,17 @@ media queries using `&` to reference the current element:
 ```tsx
 <button
   mix={css({
-    color: 'white',
-    backgroundColor: 'blue',
-    padding: '12px 24px',
-    borderRadius: '4px',
-    border: 'none',
-    cursor: 'pointer',
-    '&:hover': { backgroundColor: 'darkblue' },
-    '&:active': { transform: 'scale(0.98)' },
-    '&:disabled': { opacity: 0.5, cursor: 'not-allowed' },
-    '& .title': { fontSize: '20px', fontWeight: 'bold' },
-    '@media (max-width: 768px)': { width: '100%' },
+    color: "white",
+    backgroundColor: "blue",
+    padding: "12px 24px",
+    borderRadius: "4px",
+    border: "none",
+    cursor: "pointer",
+    "&:hover": { backgroundColor: "darkblue" },
+    "&:active": { transform: "scale(0.98)" },
+    "&:disabled": { opacity: 0.5, cursor: "not-allowed" },
+    "& .title": { fontSize: "20px", fontWeight: "bold" },
+    "@media (max-width: 768px)": { width: "100%" },
   })}
 />
 ```
@@ -87,8 +87,8 @@ state in JavaScript:
 ```tsx
 <div
   mix={css({
-    backgroundColor: 'blue', // static
-    '&:hover': { '& .title': { color: 'blue' } }, // parent hover → child
+    backgroundColor: "blue", // static
+    "&:hover": { "& .title": { color: "blue" } }, // parent hover → child
   })}
   style={{ width: `${progress}%` }} // dynamic
 />
@@ -120,7 +120,7 @@ Adds client-side navigation behavior to any element. Makes non-anchor elements b
 navigation links:
 
 ```tsx
-<article mix={link('/courses/intro')}>
+<article mix={link("/courses/intro")}>
   <h3>Introduction</h3>
 </article>
 ```
@@ -134,7 +134,7 @@ Use native DOM events directly with `on(...)`. For buttons and links, `click` al
 keyboard activation when the element has the right semantics:
 
 ```tsx
-<button mix={on('click', () => doAction())}>Action</button>
+<button mix={on("click", () => doAction())}>Action</button>
 ```
 
 For gesture-specific behavior, compose the pointer or keyboard events the interaction actually
@@ -172,7 +172,13 @@ Sets HTML attributes through the mixin system.
 Animates an element when it is inserted into the DOM. Config specifies the **starting** style:
 
 ```tsx
-<div mix={animateEntrance({ opacity: 0, transform: 'translateY(8px)', duration: 180 })} />
+<div
+  mix={animateEntrance({
+    opacity: 0,
+    transform: "translateY(8px)",
+    duration: 180,
+  })}
+/>
 ```
 
 ### `animateExit(config)`
@@ -186,11 +192,15 @@ in the DOM until the animation completes:
     <div
       key="panel"
       mix={[
-        animateEntrance({ opacity: 0, transform: 'scale(0.98)', ...spring('smooth') }),
-        animateExit({ opacity: 0, duration: 120, easing: 'ease-in' }),
+        animateEntrance({
+          opacity: 0,
+          transform: "scale(0.98)",
+          ...spring("smooth"),
+        }),
+        animateExit({ opacity: 0, duration: 120, easing: "ease-in" }),
       ]}
     />
-  )
+  );
 }
 ```
 
@@ -201,8 +211,11 @@ Animates layout changes (position/size) using FLIP-style transforms:
 ```tsx
 {
   items.map((item) => (
-    <li key={item.id} mix={animateLayout({ duration: 220, easing: 'ease-out' })} />
-  ))
+    <li
+      key={item.id}
+      mix={animateLayout({ duration: 220, easing: "ease-out" })}
+    />
+  ));
 }
 ```
