+++ app/assets/jam/2025/newsletter-subscribe.tsx
@@ -12,7 +12,7 @@ export let JamNewsletterSubscribeForm = clientEntry(
 
     return (props: { class?: string }) => (
       <form
-        action={routes.actions.newsletter.href()}
+        action={routes.newsletterSubscribe.href()}
         method="post"
         class={props.class}
         mix={[
